+++ src/main/java/Algorithm/Graph.java
@@ -11,13 +11,13 @@ public interface Graph {
 	 * @param j 
 	 * @return the cost of arc (i,j) if (i,j) is an arc; -1 otherwise
 	 */
-	public abstract int getCost(int i, int j);
+	public abstract double getCost(long i, long j);
 
 	/**
 	 * @param i 
 	 * @param j 
 	 * @return true if <code>(i,j)</code> is an arc of <code>this</code>
 	 */
-	public abstract boolean isArc(int i, int j);
+	public abstract boolean isArc(long i, long j);
 
 }
